Update README to render HTML
Removed the code fences so that the HTML is rendered, hopefully, by GitHub rather than as fenced code that is not interpreted by the browser or GitHub's Markdown rendering engine.
Update REAMDE

Shorten the border of the `bat` output so that the view is not scrolled on GitHub.
Update README


Update README


Update README


Update README


Update README


Update README


Update README


Update README


Update README

Remove code fences from the HTML blocks and shorten the table border that `bat` creates for pretty-printing so that the view isn't horizontally-scrolled on GitHub.