Update `README.md` and track all other files again
The local checkout of the Git repository was copied from an
installation of Fedora to openSUSE. This caused all the files to recieve
a new creation or modification time, so Git is interpreting this as a
real change and so the files must be added to the commit for safety,
though the files haven't actually changed in most cases.

The `README.md` and `*.slim` files were updated.

The SLiM script's output lines where changed to use `paste()` rather
than `+` string concatenation because it improves readability and
robustness.